SAA-552: Moving all movement list functions to internal location service
